+++ mongod/check.go
@@ -515,7 +515,7 @@ func (c *appRepository) processApps(cur *mongo.Cursor, ctx context.Context) ([]*
 	for cur.Next(ctx) {
 		var tempApp model.SpecificAppWithoutIDs
 		if err := cur.Decode(&tempApp); err != nil {
-			logrus.Fatal(err)
+			logrus.Error(err)
 			return nil, err
 		}
 		app := &model.SpecificAppWithoutIDs{
@@ -533,5 +533,5 @@ func (c *appRepository) processApps(cur *mongo.Cursor, ctx context.Context) ([]*
 		}
 		apps = append(apps, app)
 	}
-	return apps, nil
+	return apps, cur.Err()
 }