+++ server/utils/owner.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/gin-gonic/gin"
 	"go.mongodb.org/mongo-driver/bson"
+	"go.mongodb.org/mongo-driver/bson/primitive"
 	"go.mongodb.org/mongo-driver/mongo"
 )
 
@@ -48,6 +49,40 @@ func ResolveOwnerMiddleware(database *mongo.Database) gin.HandlerFunc {
 		c.Set(ownerContextKey, owner)
 		c.Next()
 	}
+}
+
+// EnsureTeamUserAppAccess verifies that a team user may act on the given app.
+func EnsureTeamUserAppAccess(ctx context.Context, username, appName string, database *mongo.Database) error {
+	var teamUser model.TeamUser
+	err := database.Collection("team_users").FindOne(ctx, bson.M{"username": username}).Decode(&teamUser)
+	if errors.Is(err, mongo.ErrNoDocuments) {
+		return nil
+	}
+	if err != nil {
+		return err
+	}
+
+	var appMeta struct {
+		ID primitive.ObjectID `bson:"_id"`
+	}
+	err = database.Collection("apps_meta").FindOne(ctx, bson.M{
+		"app_name": appName,
+		"owner":    teamUser.Owner,
+	}).Decode(&appMeta)
+	if err != nil {
+		if errors.Is(err, mongo.ErrNoDocuments) {
+			return errors.New("you don't have access to this app")
+		}
+		return err
+	}
+
+	for _, allowedAppID := range teamUser.Permissions.Apps.Allowed {
+		if allowedAppID == appMeta.ID.Hex() {
+			return nil
+		}
+	}
+
+	return errors.New("you don't have access to this app")
 }
 
 func GetOwnerFromContext(c *gin.Context) (string, error) {